AbstractComputationService
